Updates gradle files. Refactors seatch bar and home screen
