+++ build.gradle
@@ -1,30 +1,26 @@
 buildscript {
     ext {
         compose_ui_version = '1.4.2'
+        nav_version = '2.5.3'
+        desugar_version = '1.2.2'
+        core_version = '1.10.0'
+        lifecycle_version = '2.6.1'
+        activity_version = '1.7.1'
+        material_version = '1.4.2'
         lottie_version = '4.2.0'
-        ktx_version = '1.7.0'
         customview_version = '1.2.0-alpha02'
         customview_pool_version = '1.0.0'
-        lifecycle_runtime_ktx_version = '2.3.1'
-        activity_compose_version = '1.3.1'
-        material_version = '1.4.2'
-        nav_version = '2.5.3'
-        junit_version = '4.13.2'
-        junit_test_version = '1.1.5'
-        espresso_core_version = '3.5.1'
-        pager_version = '0.18.0'
         retrofit_version = '2.9.0'
         okhttp_version = '5.0.0-alpha.3'
         corutines_core_version = '1.4.2'
         corutines_android_version = '1.4.1'
         lifecycle_viewmodel_version = '2.2.0'
-        hilt_version = '2.40.5'
-        hilt_lifecycle_version = '1.0.0-alpha03'
+        hilt_version = '2.44'
         hilt_compiler_version = '1.0.0'
         coil_version = '2.0.0-rc01'
         share_pref_version = '1.2.0'
         gson_version = '2.8.6'
-        desugar_version = '1.2.2'
+        kotlin_version = '1.8.10'
     }
     repositories {
         google()
@@ -32,10 +28,13 @@ buildscript {
     }
     dependencies {
         classpath "com.google.dagger:hilt-android-gradle-plugin:2.40.5"
+        classpath 'com.android.tools.build:gradle:3.2.1'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }// Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
     id 'com.android.application' version '7.3.1' apply false
     id 'com.android.library' version '7.3.1' apply false
     id 'org.jetbrains.kotlin.android' version '1.6.10' apply false
+    id 'com.google.dagger.hilt.android' version '2.44' apply false
 }
